massagehandler

+++ src/main/java/com/memefest/Services/Impl/CategoryService.java
@@ -8,8 +8,9 @@ import java.util.stream.Stream;
 
 import com.memefest.DataAccess.Category;
 import com.memefest.DataAccess.CategoryFollower;
-import com.memefest.DataAccess.MainCategory;
 import com.memefest.DataAccess.SubCategory;
+import com.memefest.DataAccess.SubCategoryId;
+import com.memefest.DataAccess.TopicCategory;
 import com.memefest.DataAccess.User;
 import com.memefest.DataAccess.JSON.CategoryJSON;
 import com.memefest.DataAccess.JSON.SubCategoryJSON;
@@ -42,16 +43,12 @@ public class CategoryService implements CategoryOperations{
   //throw a custom exception to show object was not created
   public void createCategory(CategoryJSON category) {
     try{
-      getCategoryEntity(category); 
+      Category catEntity = getCategoryEntity(category); 
+    }
+    catch(NoResultException ex){
       Category newCategory = new Category();
       newCategory.setCat_Name(category.getCategoryName());
       this.entityManager.persist(newCategory);
-      createCategoryFollowers(category);
-      for (TopicJSON topic : category.getTopics())
-        topicOps.createTopic(topic); 
-    }
-    catch(NoResultException ex){
-        return;
     }
   } 
       
@@ -83,25 +80,29 @@ public class CategoryService implements CategoryOperations{
     } 
     catch(NoResultException ex){
       createCategory(category);
-      return;
     }
-    for (TopicJSON topic : category.getTopics())
-        topicOps.editTopic(topic); 
-    removeCategoryFollowers(category);
+    if(category.getTopics() != null && !category.getTopics().isEmpty())
+        for (TopicJSON topic : category.getTopics())
+          topicOps.createTopic(topic);
+    removeCategoryFollowers(category); 
     createCategoryFollowers(category);
+    
   }
     
   //throw a custom exception to show object was not created
   public void createCategoryFollowers(CategoryJSON category) {
+    if(category.getFollowedBy() == null || category.getFollowedBy().isEmpty())
+      return;
     try{
       Category foundCategory = getCategoryEntity(category);
       Set<User> catUsers = new HashSet<User>();
       try{
+
         Set<CategoryFollower> catFollowers = getCategoryFollowers(category);
         catUsers = (Set<User>)catFollowers.stream().map(catFollower -> catFollower.getUser()).collect(Collectors.toSet()); 
       }
       catch(NoResultException ex){
-       return; 
+        return;
       }
       for (UserJSON user : category.getFollowedBy()) {
         User follower = this.userOperations.getUserEntity(user);
@@ -121,49 +122,69 @@ public class CategoryService implements CategoryOperations{
   }
       
   public Set<CategoryFollower> getCategoryFollowers(CategoryJSON category) throws NoResultException{
-    Category foundCategory = getCategoryEntity(category);
-    Stream<CategoryFollower> query = this.entityManager.createNamedQuery("CategoryFollower.findByCategoryId", CategoryFollower.class)
+    try{
+      Category foundCategory = getCategoryEntity(category);
+      Stream<CategoryFollower> query = this.entityManager.createNamedQuery("CategoryFollower.findByCategoryId", CategoryFollower.class)
                     .setParameter("categoryId", Integer.valueOf(foundCategory.getCat_Id())).getResultStream();
-    return (Set<CategoryFollower>)query.map(object -> (CategoryFollower)object).collect(Collectors.toSet());
+      return (Set<CategoryFollower>)query.map(object -> (CategoryFollower)object).collect(Collectors.toSet());
+    }
+    catch(NoResultException ex){
+      throw new NoResultException();
+    }
   }
       
   public void removeCategoryFollowers(CategoryJSON category) {
+    if(category.getCancelFollowedBy() == null || category.getCancelFollowedBy().isEmpty())
+      return;
+    try{
     Set<CategoryFollower> categoryFollowers = getCategoryFollowers(category);
+    
     for (Iterator<UserJSON> iterator = category.getCancelFollowedBy().iterator(); iterator.hasNext(); ) {
-      UserJSON user = iterator.next();
-      categoryFollowers.stream().filter(candidate -> {
-          if (candidate.getUserId() == user.getUserId())
-              return true; 
-          User userEntity = this.userOperations.getUserEntity(user);
-          return (userEntity == candidate.getUser());
-        }).forEach(candidate -> this.entityManager.remove(candidate));
+        UserJSON user = iterator.next();
+        categoryFollowers.stream().filter(candidate -> {
+            if (candidate.getUserId() == user.getUserId())
+                return true; 
+            User userEntity = this.userOperations.getUserEntity(user);
+            return (userEntity == candidate.getUser());
+          }).forEach(candidate -> this.entityManager.remove(candidate));
+      }
     } 
+    catch(NoResultException ex){
+      
+    }
   }
 
       
   public Set<UserJSON> getCategoryFollowersInfo(CategoryJSON category) throws NoResultException{
-    Set<UserJSON> followers = getCategoryFollowers(category).stream().map(categFollowerEntity ->{
-      UserJSON user = new UserJSON(categFollowerEntity.getUserId(), categFollowerEntity.getUser().getUsername());
-      return user;
-    }).collect(Collectors.toSet());
-    return followers;
+    try{
+      Set<UserJSON> followers = getCategoryFollowers(category).stream().map(categFollowerEntity ->{
+        UserJSON user = new UserJSON(categFollowerEntity.getUserId(), categFollowerEntity.getUser().getUsername());
+        return user;
+      }).collect(Collectors.toSet());
+      return followers;
+    }
+    catch(NoResultException ex){
+      throw new NoResultException();
+    }
   }
       
   public Category getCategoryEntity(CategoryJSON category) throws NoResultException {
     Category foundCategory = null;
-    if((category != null) && category.getCategoryId() != 0 && category.getCategoryId() != 1) {
+    if((category != null) && category.getCategoryId() != 0 ) {
       foundCategory = (Category)this.entityManager.find(Category.class, Integer.valueOf(category.getCategoryId()));
       if (foundCategory == null){
-        return foundCategory;
+        throw new NoResultException();
       }          
-    } 
-    Query query = this.entityManager.createNamedQuery("Category.getCategoryByTitle");
-    query.setParameter("title", "%" + category.getCategoryName() + "%");
-    foundCategory = (Category)query.getSingleResult();
-    return foundCategory;
+    }
+    else if(category.getCategoryName() != null){   
+        foundCategory = this.entityManager.createNamedQuery("Category.getCategoryByTitle", Category.class)
+                                .setParameter(1,category.getCategoryName()).getSingleResult();
+    }
+    else throw new NoResultException();
+    return foundCategory; 
   }
   
-
+  /* 
   public MainCategory getMainCategoryEntity(CategoryJSON category) throws NoResultException {
     MainCategory foundCategory = null;
     if ((category != null) && category.getCategoryId() != 0 && category.getCategoryId() != 1) {
@@ -171,23 +192,85 @@ public class CategoryService implements CategoryOperations{
       if (foundCategory != null)
         return foundCategory;
     } 
-    Query query = this.entityManager.createNamedQuery("MainCategory.getCategoryByTitle");
-    query.setParameter("title", "%" + category.getCategoryName() + "%");
+    Query query = this.entityManager.createNativeQuery("MainCategory.getCategoryByTitle");
+    query.setParameter("title",category.getCategoryName());
     foundCategory = (MainCategory)query.getSingleResult();
     return foundCategory;
   }
-    
+  */
+  
+  public void editSubCategory(SubCategoryJSON subCategory){
+    /*MainCategory mainCategory = null;
+    try{
+      mainCategory = getMainCategoryFromCategory(subCategory);
+    }
+    catch(NoResultException ex){
+      createMainCategory(subCategory.getMainCategoryJSON());
+      editSubCategory(subCategory);
+    }
+    */
+    editCategory(subCategory);
+    Category categoryEntity = getCategoryEntity(subCategory);
+
+    for(CategoryJSON candidate : subCategory.getParentCategories()){
+      Category candidateEntity = null;
+      try{
+        editCategory(candidate);
+        candidateEntity = getCategoryEntity(candidate);
+      }
+      catch(NoResultException ex){
+        continue;
+      }
+      try{
+       getSubCategoryEntity(categoryEntity, candidateEntity);
+      }
+      catch(NoResultException ex){  
+        createParentCategories(categoryEntity, candidateEntity);
+      }
+    }
+    for(CategoryJSON candidate : subCategory.getCanceledParentCategories()){
+      try{
+          Category candidateEntity = getCategoryEntity(candidate);
+          SubCategory subCatEntity = getSubCategoryEntity(categoryEntity, candidateEntity);
+          entityManager.remove(subCatEntity);
+      } 
+      catch(NoResultException ex){
+        continue;
+      }
+    }
+  }
+
+  private void createParentCategories(Category category, Category parentCategory){  
+      SubCategory subCat= new SubCategory();
+      subCat.setCategory(category);
+      subCat.setParent(parentCategory);
+      entityManager.persist(subCat);
+  }
+
+  private SubCategory getSubCategoryEntity(Category category, Category parentCategory) throws NoResultException{
+    SubCategoryId subCategoryId = new SubCategoryId();
+    subCategoryId.setCat_Id(category.getCat_Id());
+    subCategoryId.setParent_Id(parentCategory.getCat_Id());
+    return entityManager.find(SubCategory.class, subCategoryId);
+  }
+/*
   //throw a custom exception to show object was not created    
-  public void createSubCategory(SubCategoryJSON category) throws IllegalArgumentException {
+  public void createSubCategory(SubCategoryJSON category) throws NoResultException{
+
+
+
+    
+
+ 
     if (category == null || (category.getCategoryName() == null && category.getCategoryId() == 0))
       return; 
-    SubCategory subCategory = null;
+    
     try {
-      subCategory = getSubCategoryEntity((CategoryJSON)category);
-      if (subCategory != null)
-        return; 
-    } catch (NoResultException ex) {
-      subCategory = new SubCategory();
+     Category categoryEntity = getCategoryEntity((CategoryJSON)category);
+      /*
+      if (category.getParentCategories() == null || category.getParentCategories().size() == 0) {
+        Set<CategoryJSON> mainCatList = (Set<CategoryJSON>)category.getParentCategories().stream().filter(candidate -> {
+              try {subCategory = new SubCategory();
       subCategory.setCat_Name(category.getCategoryName());
       if (category.getParentCategories() == null || category.getParentCategories().size() == 0) {
         Set<CategoryJSON> mainCatList = (Set<CategoryJSON>)category.getParentCategories().stream().filter(candidate -> {
@@ -212,7 +295,7 @@ public class CategoryService implements CategoryOperations{
                 throw new IllegalArgumentException("No Result Found for Category: " + candidate.getCategoryName());
               } 
             }).filter(candidate -> (candidate != null)).collect(Collectors.toSet());
-        subCategory.setParentCategories(parentcategories.stream().filter(candidate ->{
+        subCategory.setSubcategories(null); setParentCategories(parentcategories.stream().filter(candidate ->{
            return (candidate instanceof SubCategory);
         }).map(candidateEntity ->{
           try{
@@ -224,8 +307,46 @@ public class CategoryService implements CategoryOperations{
           return null;
         }).collect(Collectors.toSet()));
       } 
-      this.entityManager.persist(subCategory);
-    } 
+              } 
+            }).collect(Collectors.toSet());
+        if (mainCatList.size() == 0 || mainCatList.size() > 1)
+          return; 
+        for (CategoryJSON catInst : mainCatList) {
+          try{
+            MainCategory mainCategory = getMainCategoryFromCategory(catInst);
+            SubCategory subCategory = getSubCategoryEntity(catInst);
+
+            mainCategory.setSubcategories();
+            subCat.setMainCategory(mainCategory);
+          catch(NoResultException ex){
+
+          } 
+        } 
+        Set<Category> parentcategories = (Set<Category>)category.getParentCategories().stream().map(candidate -> {
+              try {
+                return getCategoryEntity(candidate);
+              } catch (NoResultException noResEx) {
+                throw new IllegalArgumentException("No Result Found for Category: " + candidate.getCategoryName());
+              } 
+            }).filter(candidate -> (candidate != null)).collect(Collectors.toSet());
+        subCategory.setSubcategories(null); setParentCategories(parentcategories.stream().filter(candidate ->{
+           return (candidate instanceof SubCategory);
+        }).map(candidateEntity ->{
+          try{
+            return entityManager.find(SubCategory.class, candidateEntity.getCat_Id());
+          }
+          catch(NoResultException exp){
+            ex.printStackTrace();
+          }
+          return null;
+        }).collect(Collectors.toSet()));
+      } 
+      this.entityManager.persist(subCategory);   
+    } catch (NoResultException ex) {
+      createCategory(category);
+      
+    }
+    
   }
   
   //throw a custom exception to show object was not created
@@ -265,20 +386,23 @@ public class CategoryService implements CategoryOperations{
       return;
     }
   }
+    */
       
-  public SubCategory getSubCategoryEntity(CategoryJSON category) throws NoResultException{
+  public SubCategory getSubCategoryEntity(CategoryJSON category, CategoryJSON parent) throws NoResultException{
     SubCategory foundCategory = null;
-    if ((category != null) && category.getCategoryId() != 0 && category.getCategoryId() != 1) {
-      foundCategory = (SubCategory)this.entityManager.find(SubCategory.class, Integer.valueOf(category.getCategoryId()));
+    if (((category != null) && category.getCategoryId() != 0) || 
+            ((parent != null)  && parent.getCategoryId() != 0)) {
+      SubCategoryId subCatId = new SubCategoryId();
+      subCatId.setCat_Id(category.getCategoryId());
+      subCatId.setParent_Id(parent.getCategoryId());
+      foundCategory = (SubCategory)this.entityManager.find(SubCategory.class, subCatId);
       if(foundCategory != null)
         return foundCategory;
-    } 
-    Query query = this.entityManager.createNamedQuery("SubCategoryEntity.getCategoryByTitle");
-    query.setParameter("title", "%" + category.getCategoryName() + "%");
-    foundCategory = (SubCategory)query.getSingleResult();
-    return foundCategory;
+      else throw new NoResultException();
+    }
+    else throw new NoResultException();
   }
-      
+  /* 
   //throw a custom exception to show object was not created
   public void createMainCategory(CategoryJSON category) {
     if (category == null || (category.getCategoryName() == null && category.getCategoryId() == 0))
@@ -333,7 +457,7 @@ public class CategoryService implements CategoryOperations{
       return null;
     } 
   }
-      
+  */
   public void removeCategory(CategoryJSON category) {
     if (category == null)
       return; 
@@ -344,39 +468,75 @@ public class CategoryService implements CategoryOperations{
       this.entityManager.remove(foundCategory);
     } 
   }
-      
-    
-      
+/*  
   public CategoryJSON getMainCategoryInfo(CategoryJSON mainCat) throws NoResultException {
     MainCategory mainCategoryEntity = getMainCategoryEntity(mainCat);
     CategoryJSON mainCategory = new CategoryJSON(mainCategoryEntity.getCat_Id(), mainCategoryEntity.getCat_Name(),
                                         null, null, null);
     return mainCategory;
   }
-      
-  public Set<CategoryJSON> getSubCategoryInfo(CategoryJSON category) throws NoResultException{
+*/
+  public SubCategoryJSON getSubCategoryInfo(SubCategoryJSON category) throws NoResultException{
+    CategoryJSON categoryInfo = getCategoryInfo(category);
+    Set<CategoryJSON> parentCategories = new HashSet<CategoryJSON>();
+    try{
+      Stream<SubCategory> subCats =  entityManager.createNamedQuery("SubCategory.getParentCategories",SubCategory.class)
+          .setParameter("parentId", categoryInfo.getCategoryId()).getResultStream();
+      parentCategories = subCats.map(candidate ->{
+      return getCategoryInfo(new CategoryJSON(candidate.getParent().getCat_Id(), null, null, null, null));
+              }).collect(Collectors.toSet());
+    }
+    catch(NoResultException ex){
+
+    }
+    SubCategoryJSON subCategoryJSON = new SubCategoryJSON(0, null, null, null, parentCategories);
+    subCategoryJSON.setFollowedBy(categoryInfo.getFollowedBy());
+    subCategoryJSON.setCategoryName(categoryInfo.getCategoryName());
+    subCategoryJSON.setCategoryId(categoryInfo.getCategoryId()); 
+    return subCategoryJSON;
+    /*
     SubCategory subCatEntity = getSubCategoryEntity(category);
     return (Set<CategoryJSON>)subCatEntity.getParentCategories().stream().map(candidate -> 
                 new CategoryJSON(candidate.getCat_Id(), candidate.getCat_Name(), null, null, 
                       null))        
           .collect(Collectors.toSet());
+    */
   }
     
   public Set<TopicJSON> getCategoryTopics(CategoryJSON category)throws NoResultException{
-    Category categoryEntity= getCategoryEntity(category);
+    Category catEntity = getCategoryEntity(category);
+    Set<TopicJSON> topics = new HashSet<TopicJSON>();
+    Stream<TopicCategory> topicCats = entityManager.createNamedQuery("TopicCategory.getByCategoryId", TopicCategory.class)
+              .setParameter("categoryId", catEntity.getCat_Id()).getResultStream();
+    topics = topicCats.map(candidate ->{
+      return new TopicJSON(candidate.getTopic_Id() , null, null, null, null, null);
+    }).collect(Collectors.toSet());
+    return topics;
+    /*Category categoryEntity= getCategoryEntity(category);
     return categoryEntity.getTopics().stream().map(topicEntity ->{
       TopicJSON topicJson = new TopicJSON(topicEntity.getTopic_Id(), topicEntity.getTitle(), null, null, null, null);
       topicJson.setTopicId(topicEntity.getTopic_Id());
       topicJson.setTitle(topicEntity.getTitle());
       return topicJson;
     }).collect(Collectors.toSet());
+    */
   }
       
   public CategoryJSON getCategoryInfo(CategoryJSON category) throws NoResultException{
-    Category categoryEntity = getCategoryEntity(category);
-    Set<TopicJSON> categoryTopics = getCategoryTopics(category);
-    Set<UserJSON> followers = getCategoryFollowersInfo(category);
-    CategoryJSON categoryJson = new CategoryJSON(categoryEntity.getCat_Id(), categoryEntity.getCat_Name(), categoryTopics, followers, null);
-    return categoryJson;
-  }
+      Category categoryEntity = getCategoryEntity(category);
+      Set<TopicJSON> categoryTopics = null;
+      Set<UserJSON> followers = null;
+      try{
+        followers = getCategoryFollowersInfo(category);
+      }
+      catch(NoResultException ex){  
+      }
+      try{
+        categoryTopics = getCategoryTopics(category);
+      }
+      catch(NoResultException ex){
+      }
+      CategoryJSON categoryJson = new CategoryJSON(categoryEntity.getCat_Id(), categoryEntity.getCat_Name(), categoryTopics, followers, null);
+      return categoryJson;
+    }
 }
